Feature: 팀 정보에 팀장 정보 추가 (#101)
* feat: 팀원 조회를 위한 TeamUserRepository 쿼리 추가 작성

* feat: TeamUserJpaEntity에 isLeader 속성 추가 작성

* feat: 특정 팀에 리더를 설정하는 setLeader 쿼리 작성

* feat: 팀 정보 DTO에 팀장 정보 추가 작성

* refact: 비즈니스 규칙과 연관된 코드 이동 (응용계층 -> 도메인계층)

* feat: Team 엔티티의 setLeader 메소드 작성

* feat: 팀 생성시 팀장 설정하는 기능 작성

* feat: TeamService의 DTO 생성시 팀장 정보 반영하는 로직 작성

* fix: lazy loading에 의한 데이터 조회 문제 수정

* feat: 특정 유저가 팀장인지 확인하는 로직 추가

* feat: 팀장을 내보낼 수 없는 ErrorType 정의

* feat: 팀장은 팀에서 퇴출할 수 없는 규칙 작성

* feat: 실제 필요한 정보를 고려해 TeamResponseDto 형식 변경

+++ tlog-was/src/main/java/com/se/Tlog/domain/Team/domain/Team.java
@@ -58,10 +58,17 @@ public class Team {
 		if (repoService.isExistInTeam(this.getId(), user.getId()))
 			throw new CustomException(ErrorType.ALREADY_EXIST_IN_TEAM);
 		
-		repoService.addUserToTeam(id, user.getId());
+		repoService.addUserToTeam(this.id, user.getId());
 		// 기타 팀원 추가시 처리내용
 		
 		log.info("팀원을 팀 " + this.name + "에 추가합니다. : " + user.getName());
+	}
+	
+	public void setLeader(User user, TeamRepositoryService repoService) {
+	    if (!repoService.isExistInTeam(this.id, user.getId()))
+            throw new CustomException(ErrorType.TEAM_USER_NOT_FOUND);
+	    
+	    repoService.setLeader(this.id, user.getId());
 	}
 	
 	public void deleteUser(User user, TeamRepositoryService repoService) {
@@ -69,6 +76,8 @@ public class Team {
 			throw new CustomException(ErrorType.TEAM_USER_NOT_FOUND);
 		if (repoService.countMemberInTeam(id) == 1)
 			throw new CustomException(ErrorType.TEAM_CANNOT_BE_ORPHAN);
+		if (repoService.isLeader(id, user.getId()))
+		    throw new CustomException(ErrorType.CANNOT_REMOVE_TEAM_LEADER);
 		
 		repoService.deleteUserInTeam(this.id, user.getId());
 		// 기타 팀원 삭제 후 처리내용
